Bug 960653 - part 2 - change JSJitInfo.{protoID,depth} to 16-bit fields; r=bz,efaust

+++ js/src/jsfriendapi.h
@@ -1517,8 +1517,8 @@ struct JSJitInfo {
         JSParallelNative parallelNative;
     };
 
-    uint32_t protoID;
-    uint32_t depth;
+    uint16_t protoID;
+    uint16_t depth;
     // type not being ParallelNative means this is a DOM method.  If you
     // change that, come up with a different way of implementing
     // isDOMJitInfo().
